Name all local functions and variables 'glusterfs' instead of 'gluster'
All local functions are now prefixed with glusterfs_ and functions
available in packet-gluster.h have gluster_ as prefix.

Also move gluster_rpc_dissect_dict() from packet-gluster.c to the
GlusterFS dissector as it is the main user of the function.

Drop the unused "GlusterFS Mops" tables and protocol pointers. GlusterFS
3.2.x and newer do not seem to use this anymore.

Signed-off-by: Niels de Vos <ndevos@redhat.com>